+++ R/app_ui.R
@@ -51,61 +51,15 @@ app_ui = function(request) {
           # Right hand side - outputs ------------------------------------------
           shiny::column(6,
             # choose cluster id
-            shiny::selectInput(inputId = "cluster_id",
-                               label = "Choose a cluster id:",
-                               choices = get_all_clusters(filename = system.file("app", "www", "data",
-                                                                                 "scanner_output",
-                                                                                 package = "tfpbrowser",
-                                                                                 mustWork = TRUE))),
+            clusterid_ui("choice1"),
+
             # output options
             shiny::tabsetPanel(id = "plot_tabs",
              # Tables tab
-             shiny::tabPanel("Tables",
-                             # drop down menu to select table
-                             shiny::br(),
-                             shiny::uiOutput("choose_table"),
-                             # display table
-                             shiny::wellPanel(
-                               shiny::fluidRow(shiny::column(
-                                 12,
-                                 reactable::reactableOutput("display_table"),
-                                 align = "center",
-                                 style = "height:400px;"
-                               )),
-                               style = "background: white"
-                             ),
-                             # download button to download current table
-                             shiny::br(),
-                             shiny::fluidRow(
-                               shiny::column(12,
-                                             align = "center",
-                                             shiny::uiOutput("download_table_button") # nolint
-                               )
-                             )
-             ),
+             tablesUI("table1"),
              # Plots tab
-             shiny::tabPanel("Plots",
-                             # drop down menu to select plot
-                             shiny::br(),
-                             shiny::uiOutput("choose_plot"),
-                             # display plot
-                             shiny::wellPanel(
-                               shiny::fluidRow(shiny::column(
-                                 12,
-                                 align = "center",
-                                 shiny::uiOutput("display_plot"),
-                                 style = "height:400px;")),
-                               style = "background: white"
-                             ),
-                             # download button to download current plot
-                             shiny::br(),
-                             shiny::fluidRow(
-                               shiny::column(12,
-                                             align = "center",
-                                             shiny::uiOutput("download_plot_button") #nolint
-                               )
-                             )
-             )
+             plotsUI("plot1")
+
            )
           ) # end right column
         ) # end fluid row